fix: make event propagation work correctly
